+++ src_cpp/NBC/NBC_suite.cpp
@@ -49,7 +49,7 @@ goods and services.
 */
 
 #if defined MPI1 || defined RMA || defined MPIIO || defined EXT
-#error Legacy benchmark components can't be linked together
+#error Legacy benchmark components cant be linked together
 #endif
 
 #include <set>
@@ -488,7 +488,7 @@ template <> bool BenchmarkSuite<BS_NBC>::prepare(const args_parser &parser, cons
     }
 
     int alignment = parser.get<int>("alignment");
-    if (alignment < sizeof(void*)) {
+    if (alignment < (int)sizeof(void*)) {
         alignment = sizeof(void*);
     }
     int power2 = 1;
